memory, changingloop

+++ assets/js/script.js
@@ -48,10 +48,7 @@ function colorChanges(){
 };
 
 $(document).ready(function(){
-    // TODO: Add a listener for click events on the save button. This code should
-    // use the id in the containing time-block as a key to save the user input in
-    // local storage. HINT: What does `this` reference in the click listener
-    // function? 
+    memory();
 
     colorChanges();
 
@@ -70,55 +67,53 @@ $(document).ready(function(){
         });
 });
 
-   
-
 function saveTextArea(textContent, dataHour){ 
-    console.log(textContent, dataHour)
+    console.log(textContent, dataHour);
 
     let tasks = JSON.parse(localStorage.getItem("tasks")) || [];  
 
+   
     let existingTask = tasks.find(task => task.blockHour === dataHour);
 
     if(existingTask) {
-        existingTask.textContent = textContent;
+
+        if(existingTask.textContent !== textContent){
+            existingTask.textContent = textContent;
+            
+          
+            localStorage.setItem("tasks", JSON.stringify(tasks));
+        }
     } else {
+      
         let newTask = {blockHour: dataHour, textContent: textContent};
+    
         tasks.push(newTask);
-    }
-
-
-    tasks.push({blockHour:dataHour, textContent:textContent})
-
-    localStorage.setItem("tasks", JSON.stringify(tasks));
-}
-
-// timeBlocksContainer.addEventListener("click", function(event){
-//     // Use `.closest()` to find the nearest ancestor which is a save button
-//    
-
-//     
-//         // Use `.closest()` to find the nearest ancestor that is a time-block
-//         // Then use `querySelector` to find the textarea within that time-block
-//         const timeBlock = saveButton.closest('.time-block');
-//         const textArea = timeBlock.querySelector('.description');
-
-//         // Log the value of the textarea
-//         console.log(textArea.value);
-//     }
-
 
     
-    //How can DOM traversal be used to get the "hour-x" id of the
-    // time-block containing the button that was clicked? How might the id be
-    // useful when saving the description in local storage?
-    //
+        localStorage.setItem("tasks", JSON.stringify(tasks));
+    }
+}
 
-    // TODO: Add code to get any user input that was saved in localStorage and set
-    // the values of the corresponding textarea elements. HINT: How can the id
-    // attribute of each time-block be used to do this?
-    //
 
-    // TODO: Add code to display the current date in the header of the page.
+function memory() {
+
+    let tasks = JSON.parse(localStorage.getItem("tasks")) || [];
+
+    let blockHourElements = document.querySelectorAll('.time-block');
+    blockHourElements.forEach(blockHourElement => {
+    
+        let dataHour = blockHourElement.getAttribute('data-hour');
+
+        let task = tasks.find(task => task.blockHour === dataHour);
+
+        if (task) {
+            let textContent = task.textContent;
+            blockHourElement.querySelector('.description').value = textContent;
+        }
+    });
+};
+
+
